feat(app): integrating whitelisting and Admin Controller with SNOW
