--- private/guids.json.example
@@ -1,1 +1,0 @@
-["8dad96cb-c3db-11e6-9c43-bc0000d20000","12345678-9abc-def1-2345-123456789000"]
